add OpenEXR include directory to include path
I talked with Ken Museth about this and he noted that he had this same problem
for older versions of Houdini and since around 12.5.5xx it has disappeared.
Since we are building without Houdini, I can only assume this must correspond
to older versions of OpenEXR. Ubuntu uses 1.6 which is a few versions behind
stable, so this could be the problem.